Add object-rsync-timeout option
Adds the rsync_timeout option to to object-server.conf, so we can adjust
it away from the default of 900s.

If there are a number of large partitions needing replication,
occasionally one needs to adjust the timeout in order to allow the rsync
to complete rather than timeout and retry.

Note: Also includes change to amulet tests for validating the workers,
which is needed for the updated worker calculation count.

Change-Id: I2d895741cb0528836a675deb6399005a5bf59ab5
Depends-On: Ia250f45b4f734f5a25c1deebbf6d7fa5303b6ad9
Closes-bug: 1702039
(cherry picked from commit 18419a87c50f798b1468e7265b30836f70eba1d1)